add object-tracking and kinematics submodules (work in progress)

+++ autonomous_roboclaw/roboclaw.py
@@ -749,10 +749,10 @@ class Roboclaw:
 		return self._write1(address,self.Cmd.SETMAXLB,val)
 
 	def SetM1VelocityPID(self,address,p,i,d,qpps):
-		return self._write4444(address,self.Cmd.SETM1PID,long(d*65536),long(p*65536),long(i*65536),qpps)
+		return self._write4444(address,self.Cmd.SETM1PID,int(d*65536),int(p*65536),int(i*65536),qpps)
 
 	def SetM2VelocityPID(self,address,p,i,d,qpps):
-		return self._write4444(address,self.Cmd.SETM2PID,long(d*65536),long(p*65536),long(i*65536),qpps)
+		return self._write4444(address,self.Cmd.SETM2PID,int(d*65536),int(p*65536),int(i*65536),qpps)
 
 	def ReadISpeedM1(self,address):
 		return self._read4_1(address,self.Cmd.GETM1ISPEED)
@@ -891,10 +891,10 @@ class Roboclaw:
 		return (0,0,0)
 
 	def SetM1PositionPID(self,address,kp,ki,kd,kimax,deadzone,min,max):
-		return self._write4444444(address,self.Cmd.SETM1POSPID,long(kd*1024),long(kp*1024),long(ki*1024),kimax,deadzone,min,max)
+		return self._write4444444(address,self.Cmd.SETM1POSPID,int(kd*1024),int(kp*1024),int(ki*1024),kimax,deadzone,min,max)
 
 	def SetM2PositionPID(self,address,kp,ki,kd,kimax,deadzone,min,max):
-		return self._write4444444(address,self.Cmd.SETM2POSPID,long(kd*1024),long(kp*1024),long(ki*1024),kimax,deadzone,min,max)
+		return self._write4444444(address,self.Cmd.SETM2POSPID,int(kd*1024),int(kp*1024),int(ki*1024),kimax,deadzone,min,max)
 
 	def ReadM1PositionPID(self,address):
 		data = self._read_n(address,self.Cmd.READM1POSPID,7)
